# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Daily Fleece API
  version: "0.1.0"
  description: |
    Spec-first REST API for the Daily Fleece daily standup quiz app.

    ## Overview

    A host creates a session by uploading two photos (one knowledge question, one
    geography question). Players join the lobby, the host starts the quiz, players
    answer both questions, and the host declares the correct answers. Scores
    accumulate on a persistent leaderboard.

    ## Actors

    - **Host** — the player who creates and controls the session. Also answers as a player.
    - **Player** — any participant, including the host.

    ## Session lifecycle

    ```
    Lobby → Active → Ended
                └── Q1 Open → Q1 Closed → Q2 Open → Q2 Closed
    ```

    Transitions are triggered by host actions:
    - `POST /sessions/{projectId}/{date}/start` — Lobby → Active (opens Q1 voting)
    - `POST /sessions/{projectId}/{date}/questions/q1/correct` — closes Q1 voting, scores Q1, opens Q2 voting
    - `POST /sessions/{projectId}/{date}/questions/q2/correct` — closes Q2 voting, scores Q2, ends session

    ## Stub authentication

    Authentication is not implemented yet. Players register with a `companyId`
    (their stable company identity, max 20 characters) and a `displayName`. The
    server performs a get-or-create keyed on `companyId`: if a player with that
    ID already exists their `playerId` is returned, otherwise a new player record
...

    ## One session per day

...

...

...
servers:
...
paths:
...
  /sessions/{projectId}/{date}:
    get:
      operationId: getSessionByKey
      tags: [Sessions]
      summary: Get session by natural key
      description: |
        Returns the session for the given project and date.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SessionDate"
      responses:
        "200":
          description: Session state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "404":
          description: Session not found.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "/problems/session-not-found"
                title: "Session Not Found"
                status: 404
                detail: "No session found for project 'default' on 2026-06-10."
    post:
      operationId: createSessionByKey
      tags: [Sessions]
      summary: Create a session by natural key
      description: |
        Creates a new session for the given project and date in the Lobby phase.
        Both photos and the host's identity are submitted in a single multipart
        request. The host is automatically added as the first player.

        Returns 409 if a session already exists for this project and date.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SessionDate"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created in Lobby phase.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "409":
          description: A session already exists for this project and date.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "/problems/session-already-exists"
                title: "Session Already Exists"
                status: 409
                detail: "A session already exists for default/2026-06-10."
    delete:
      operationId: deleteSessionByKey
      tags: [Sessions]
      summary: Delete a session by natural key
      description: |
        Deletes the session for the given project and date, including its
        GridFS photos (cascade). Idempotent: returns 204 even if no session
        exists. Used by the e2e test suite to reset state between runs.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SessionDate"
      responses:
        "204":
          description: Session deleted (or did not exist).

  # ---------------------------------------------------------------------------
  # Results
  # ---------------------------------------------------------------------------
...
  /sessions/{projectId}/{date}/join:
    post:
      operationId: joinSessionByKey
      tags: [Sessions]
      summary: Join the session lobby by natural key
      description: |
        Adds the player to the session's player list. Only possible while the
        session is in Lobby phase.

        Returns 409 if the session has already started.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SessionDate"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinSessionRequest"
            example:
              playerId: "p-4f3a9b"
              displayName: "Thomas"
      responses:
        "200":
          description: Joined successfully. Returns updated session state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "404":
          description: Session not found.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "/problems/session-not-found"
                title: "Session Not Found"
                status: 404
                detail: "No session found for project 'default' on 2026-06-10."
        "409":
          description: Session is already active; cannot join.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "/problems/session-already-active"
                title: "Session Already Active"
                status: 409
                detail: "This session is already in progress. New players cannot join."
  /sessions/{projectId}/{date}/start:
    post:
      operationId: startSessionByKey
      tags: [Sessions]
      summary: Start the session by natural key
      description: |
        Transitions the session from Lobby to Active phase and opens Q1 voting.
        Only the host may call this endpoint.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SessionDate"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostActionRequest"
            example:
              hostId: "p-4f3a9b"
      responses:
        "200":
          description: Session started. Returns updated session state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "403":
          description: Caller is not the host.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "/problems/not-the-host"
                title: "Forbidden"
                status: 403
                detail: "Only the host may start the session."
        "404":
          description: Session not found.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "/problems/session-not-found"
                title: "Session Not Found"
                status: 404
                detail: "No session found for project 'default' on 2026-06-10."
        "409":
          description: Session is not in Lobby phase.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "/problems/invalid-phase-transition"
                title: "Invalid Phase Transition"
                status: 409
                detail: "Session is in Active phase and cannot be started again."

  # ---------------------------------------------------------------------------
  # Voting
  # ---------------------------------------------------------------------------
...
  /sessions/{projectId}/{date}/photos/{question}:
    get:
      operationId: getSessionPhotoByKey
      tags: [Sessions]
      summary: Get a session photo by natural key
      description: |
        Returns the raw image bytes for the Q1 or Q2 photo uploaded when the
        session was created. Available in any session phase.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SessionDate"
        - $ref: "#/components/parameters/Question"
      responses:
        "200":
          description: Photo image bytes.
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Session not found or photo expired (28-day TTL).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "/problems/session-not-found"
                title: "Session Not Found"
                status: 404
                detail: "No session found for project 'default' on 2026-06-10."
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: The project identifier. Always "default" for now.
      schema:
        type: string
      example: "default"
    SessionDate:
      name: date
      in: path
      required: true
      description: The session date in YYYY-MM-DD format (server timezone).
      schema:
        type: string
        format: date
      example: "2026-06-10"
    Question:
      name: question
      in: path
      required: true
      description: Which question — `q1` (knowledge) or `q2` (geography).
      schema:
        type: string
        enum: [q1, q2]
  schemas:
...
    CreateSessionRequest:
      type: object
      required: [hostId, hostDisplayName, q1, q2]
      properties:
        hostId:
          type: string
          description: The playerId of the player creating the session. They become the host and are automatically added to the player list.
          example: "p-4f3a9b"
        hostDisplayName:
          type: string
          maxLength: 20
          description: Display name of the host. Denormalised for rendering without a player lookup, consistent with JoinSessionRequest.
          example: "Thomas"
        q1:
          type: string
          format: binary
          description: Photo for Q1 (knowledge question — image of the calendar page).
        q2:
          type: string
          format: binary
          description: Photo for Q2 (geography question — image of the location).
    JoinSessionRequest:
      type: object
      required: [playerId, displayName]
      properties:
        playerId:
          type: string
          description: The playerId of the joining player.
          example: "p-8b1c3a"
        displayName:
          type: string
          maxLength: 20
          description: Display name shown in the player list. Supplied by the caller to avoid a cross-module lookup.
          example: "Thomas"
    HostActionRequest:
      type: object
      required: [hostId]
      properties:
        hostId:
          type: string
          description: Must match the session's hostId. Used for authorization until real auth is implemented.
          example: "p-4f3a9b"
    SessionResponse:
      type: object
      required:
        [sessionId, projectId, date, phase, hostId, players, voting]
      description: Full session document. Returned by all session state-changing endpoints.
      properties:
        sessionId:
          type: string
          example: "s-7c2d1e"
        projectId:
          type: string
          description: Always "default" for now. Multi-project support is deferred.
          example: "default"
        date:
          type: string
          format: date
          description: Session date in YYYY-MM-DD format, server timezone.
          example: "2026-05-25"
        phase:
          type: string
          enum: [Lobby, Active, Ended]
          description: |
            Current lifecycle phase.
            - `Lobby` — session created, waiting for players and host to start
            - `Active` — quiz in progress
            - `Ended` — both questions scored, results available
          example: "Active"
        hostId:
          type: string
          description: playerId of the host.
          example: "p-4f3a9b"
        players:
          type: array
          description: All players who have joined the session, including the host.
          items:
            $ref: "#/components/schemas/SessionPlayer"
        voting:
          $ref: "#/components/schemas/VotingState"
    SessionPlayer:
      type: object
      required: [playerId, displayName]
      description: A player as embedded in the session document. Display name is denormalised here for rendering without a separate player lookup.
      properties:
        playerId:
          type: string
          example: "p-4f3a9b"
        displayName:
          type: string
          example: "Thomas"
    VotingState:
      type: object
      required: [q1, q2]
      properties:
        q1:
          $ref: "#/components/schemas/QuestionVoting"
        q2:
          $ref: "#/components/schemas/QuestionVoting"
    QuestionVoting:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Open, Closed]
          description: |
            `Open` — voting in progress, answers accepted.
            `Closed` — voting ended by host setting the correct answer.
        correctAnswer:
          type: string
          nullable: true
          description: |
            Set when the host declares the correct answer. Null while voting is open.
            A/B/C for Q1; ISO 3166-1 alpha-2 for Q2.
          example: "B"
        answerCount:
          type: integer
          description: |
            Number of players who have submitted an answer. Always present,
            regardless of voting status.
          example: 3
        answers:
          type: object
          description: |
            Map of playerId → PlayerAnswer. Hidden from all clients while voting
            is Open. Revealed to all once the correct answer is set.
          additionalProperties:
            $ref: "#/components/schemas/PlayerAnswer"
    PlayerAnswer:
      type: object
      required: [displayName, answer]
      properties:
        displayName:
          type: string
          description: Denormalised for rendering without a player lookup.
          example: "Thomas"
        answer:
          type: string
          description: A/B/C for Q1; ISO 3166-1 alpha-2 country code for Q2.
          example: "B"

    # -------------------------------------------------------------------------
    # Voting
    # -------------------------------------------------------------------------
...
    ProblemDetail:
      type: object
      description: |
        Error response following [RFC 9457 Problem Details](https://datatracker.ietf.org/doc/html/rfc9457).
      properties:
        type:
          type: string
          description: |
            A URI reference identifying the problem type. Relative URIs are
            resolved against the server base URL. Use "about:blank" when no
            additional documentation is available.
          example: "/problems/session-not-found"
        title:
          type: string
          description: Short, human-readable summary of the problem type. Stable across occurrences.
          example: "Session Not Found"
        status:
          type: integer
          description: The HTTP status code for this occurrence.
          example: 404
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence. Helps the client understand what went wrong.
          example: "No session found with ID s-7c2d1e."
        instance:
          type: string
          description: Optional URI reference identifying this specific occurrence of the problem.
          example: "/sessions/s-7c2d1e"