# backend/src/main/java/de/dailyfleece/backend/quiz/infrastructure/web/SessionController.java
package de.dailyfleece.backend.quiz.infrastructure.web;

import de.dailyfleece.api.SessionsApi;
import de.dailyfleece.api.model.HostActionRequest;
import de.dailyfleece.api.model.JoinSessionRequest;
import de.dailyfleece.api.model.SessionResponse;
import de.dailyfleece.backend.quiz.application.CreateSessionUseCase;
import de.dailyfleece.backend.quiz.application.DeleteSessionUseCase;
import de.dailyfleece.backend.quiz.application.JoinSessionUseCase;
import de.dailyfleece.backend.quiz.application.LoadSessionPhotoUseCase;
import de.dailyfleece.backend.quiz.application.LoadSessionUseCase;
import de.dailyfleece.backend.quiz.application.NoSessionForDate;
import de.dailyfleece.backend.quiz.application.StartSessionUseCase;
import de.dailyfleece.backend.quiz.domain.Photo;
import de.dailyfleece.backend.quiz.domain.PhotoType;
import de.dailyfleece.backend.quiz.domain.ProjectId;
import de.dailyfleece.backend.quiz.domain.Session;
import de.dailyfleece.backend.quiz.domain.SessionKey;
import de.dailyfleece.backend.shared.PlayerName;
import java.io.IOException;
import java.time.LocalDate;
import java.util.UUID;
import org.springframework.core.io.InputStreamResource;
import org.springframework.core.io.Resource;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

@RestController
@RequestMapping(value = "/api/{version}", version = "1.0+")
class SessionController implements SessionsApi {

    private final CreateSessionUseCase createSessionUseCase;
    private final DeleteSessionUseCase deleteSessionUseCase;
    private final LoadSessionUseCase loadSessionUseCase;
    private final LoadSessionPhotoUseCase loadSessionPhotoUseCase;
    private final JoinSessionUseCase joinSessionUseCase;
    private final StartSessionUseCase startSessionUseCase;
    private final SessionResponseMapper mapper;

    SessionController(
            CreateSessionUseCase createSessionUseCase,
            DeleteSessionUseCase deleteSessionUseCase,
            LoadSessionUseCase loadSessionUseCase,
            LoadSessionPhotoUseCase loadSessionPhotoUseCase,
            JoinSessionUseCase joinSessionUseCase,
            StartSessionUseCase startSessionUseCase,
            SessionResponseMapper mapper) {
        this.createSessionUseCase = createSessionUseCase;
        this.deleteSessionUseCase = deleteSessionUseCase;
        this.loadSessionUseCase = loadSessionUseCase;
        this.loadSessionPhotoUseCase = loadSessionPhotoUseCase;
        this.joinSessionUseCase = joinSessionUseCase;
        this.startSessionUseCase = startSessionUseCase;
        this.mapper = mapper;
    }

    @Override
    public ResponseEntity<Resource> getSessionPhotoByKey(String projectId, LocalDate date, String question) {
        Session session = loadSessionUseCase.load(date).orElseThrow(() -> new NoSessionForDate(date));
        Photo photo = loadSessionPhotoUseCase.load(session.sessionId(), question);
        MediaType contentType = PhotoTypeRegistry.toMediaType(photo.photoType());
        return ResponseEntity.ok().contentType(contentType).body(new InputStreamResource(photo.data()));
    }

    @Override
    public ResponseEntity<SessionResponse> createSessionByKey(
            String projectId,
            LocalDate date,
            String hostId,
            String hostDisplayName,
            MultipartFile q1,
            MultipartFile q2) {
        SessionKey key = new SessionKey(new ProjectId(projectId), date);
        return createSessionForKey(key, hostId, hostDisplayName, q1, q2);
    }

    @Override
    public ResponseEntity<Void> deleteSessionByKey(String projectId, LocalDate date) {
        deleteSessionUseCase.delete(date);
        return ResponseEntity.noContent().build();
    }

    @Override
    public ResponseEntity<SessionResponse> getSessionByKey(String projectId, LocalDate date) {
        return loadSessionUseCase
                .load(date)
                .map(mapper::toResponse)
                .map(ResponseEntity::ok)
                .orElseThrow(() -> new NoSessionForDate(date));
    }

    @Override
    public ResponseEntity<SessionResponse> joinSessionByKey(
            String projectId, LocalDate date, JoinSessionRequest request) {
        Session sessionForKey = loadSessionUseCase.load(date).orElseThrow(() -> new NoSessionForDate(date));
        UUID playerId = UUID.fromString(request.getPlayerId());
        PlayerName displayName = new PlayerName(request.getDisplayName());
        Session updated = joinSessionUseCase.join(sessionForKey.sessionId(), playerId, displayName);
        return ResponseEntity.ok(mapper.toResponse(updated));
    }

    @Override
    public ResponseEntity<SessionResponse> startSessionByKey(
            String projectId, LocalDate date, HostActionRequest request) {
        Session sessionForKey = loadSessionUseCase.load(date).orElseThrow(() -> new NoSessionForDate(date));
        UUID hostId = UUID.fromString(request.getHostId());
        Session updated = startSessionUseCase.start(sessionForKey.sessionId(), hostId);
        return ResponseEntity.ok(mapper.toResponse(updated));
    }

    private ResponseEntity<SessionResponse> createSessionForKey(
            SessionKey key, String hostId, String hostDisplayName, MultipartFile q1, MultipartFile q2) {
        try {
            UUID hostUuid = UUID.fromString(hostId);
            PlayerName name = new PlayerName(hostDisplayName);
            PhotoType q1Type = PhotoTypeRegistry.toEnum(q1.getContentType());
            PhotoType q2Type = PhotoTypeRegistry.toEnum(q2.getContentType());
            Session session = createSessionUseCase.create(
                    key, hostUuid, name, q1.getInputStream(), q1Type, q2.getInputStream(), q2Type);
            return ResponseEntity.status(201).body(mapper.toResponse(session));
        } catch (IOException e) {
            throw new IllegalStateException("Failed to read uploaded photo", e);
        }
    }
}